Updated Dantoonie sample
